LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/cross-window/open-samedocument-crossorigin-expected.txt



FAIL using window.open() to navigate same-document targeting a cross-origin window assert_equals: navigationType expected "push" but got "replace"
